add statements to the grammar and support `let`
The grammar now parses a sequence of statements instead of a single
expression.

The forms `let id: type = expr` and `let id = expr` are now supported
by the type checker and parser. But the identifier is not yet tracked
for later usage in the program.

Added the concept of a Void type for statements that do not return any
value and updated the compiler to compile wasm functions with no return
values accordingly.