embed type errors in the AST
Now that type errors are just nodes in the AST,
the compiler can check the AST for errors and report groups of errors
to the user in a more intuitive and compact format.

Type error recovery is also enhanced by these changes.
The parser is able to build a complete typed AST despite errors
in most cases.